+++ extension/index.html
@@ -1,13 +1,23 @@
 <!doctype html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React</title>
-  </head>
-  <body>
-    <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
-  </body>
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>BotBlocker</title>
+  <style>
+    body {
+      margin: 0;
+      padding: 0;
+    }
+    #root {
+      margin: 0;
+      padding: 0;
+      width: 100%;
+    }
+  </style>
+</head>
+<body>
+<div id="root"></div>
+<script type="module" src="/src/popUp/main.jsx"></script>
+</body>
 </html>
