+++ extension/public/manifest.json
@@ -6,5 +6,5 @@
   "action": {
     "default_popup": "index.html"
   },
-  "options_page": "options.html"
+  "options_page": "settings.html"
 }
