+++ extension/src/popUp/components/voting/QuestionnaireYes.jsx
