+++ extension/src/popUp/index.css
@@ -2,8 +2,6 @@
 body {
   width: 400px; /* Largura do popup */
   height: 600px; /* Altura do popup */
-  margin: 0; /* Remove margens padrão */
-  padding: 0; /* Remove padding padrão */
   background-color: white; /* Fundo branco */
   font-family: Arial, sans-serif; /* Fonte padrão */
   overflow: hidden; /* Evita barras de rolagem */
@@ -11,7 +9,6 @@ body {
 
 /* Estilo para o container principal */
 #root {
-  width: 100%;
   height: 100%;
   display: flex;
   flex-direction: column;
