+++ extension/src/settings/components/global/Navbar.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import Logo from '../../assets/logo.png';
-import { FaCog } from 'react-icons/fa';
+import Logo from '../../../assets/logo.png';
 
 const Nav = styled.nav`
     display: flex;
@@ -18,29 +17,30 @@ const LogoImage = styled.img`
     height: 40px;
 `;
 
-const GearIcon = styled(FaCog)`
-    font-size: 24px;
+const LoginButton = styled.button`
+    background-color: #4a5b84;
     color: white;
+    border: none;
+    padding: 8px 16px;
+    border-radius: 4px;
+    font-weight: 500;
     cursor: pointer;
+    transition: background-color 0.2s;
 
     &:hover {
-        opacity: 0.8;
+        background-color: #3a4a70;
+    }
+
+    &:active {
+        background-color: #2d3a5a;
     }
 `;
 
 const Navbar = () => {
-    const handleSettingsClick = () => {
-        chrome.tabs.create({ url: chrome.runtime.getURL('settings.html') });
-    };
-
     return (
         <Nav>
             <div>
                 <LogoImage src={Logo} alt="Logo" />
-            </div>
-
-            <div>
-                <GearIcon onClick={handleSettingsClick} />
             </div>
         </Nav>
     );
